Imported Upstream version 1.0.0~0alpha

+++ src/etc/get-snapshot.py
@@ -1,5 +1,14 @@
 #!/usr/bin/env python
-# xfail-license
+#
+# Copyright 2011-2014 The Rust Project Developers. See the COPYRIGHT
+# file at the top-level directory of this distribution and at
+# http://rust-lang.org/COPYRIGHT.
+#
+# Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+# http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+# <LICENSE-MIT or http://opensource.org/licenses/MIT>, at your
+# option. This file may not be copied, modified, or distributed
+# except according to those terms.
 
 import os, tarfile, re, shutil, sys
 from snapshot import *
@@ -8,51 +17,30 @@ def unpack_snapshot(triple, dl_path):
   print("opening snapshot " + dl_path)
   tar = tarfile.open(dl_path)
   kernel = get_kernel(triple)
+
+  stagep = os.path.join(triple, "stage0")
+
+  # Remove files from prior unpackings, since snapshot rustc may not
+  # be able to disambiguate between multiple candidate libraries.
+  # (Leave dirs in place since extracting step still needs them.)
+  for root, _, files in os.walk(stagep):
+    for f in files:
+      print("removing " + os.path.join(root, f))
+      os.unlink(os.path.join(root, f))
+
   for p in tar.getnames():
     name = p.replace("rust-stage0/", "", 1);
-    stagep = os.path.join(triple, "stage0")
+
     fp = os.path.join(stagep, name)
     print("extracting " + p)
     tar.extract(p, download_unpack_base)
     tp = os.path.join(download_unpack_base, p)
+    if os.path.isdir(tp) and os.path.exists(fp):
+        continue
     shutil.move(tp, fp)
   tar.close()
   shutil.rmtree(download_unpack_base)
 
-def determine_curr_snapshot(triple):
-  i = 0
-  platform = get_platform(triple)
-
-  found_file = False
-  found_snap = False
-  hsh = None
-  date = None
-  rev = None
-
-  f = open(snapshotfile)
-  for line in f.readlines():
-    i += 1
-    parsed = parse_line(i, line)
-    if (not parsed): continue
-
-    if found_snap and parsed["type"] == "file":
-      if parsed["platform"] == platform:
-        hsh = parsed["hash"]
-        found_file = True
-        break;
-    elif parsed["type"] == "snapshot":
-      date = parsed["date"]
-      rev = parsed["rev"]
-      found_snap = True
-
-  if not found_snap:
-    raise Exception("no snapshot entries in file")
-
-  if not found_file:
-    raise Exception("no snapshot file found for platform %s, rev %s" %
-                    (platform, rev))
-
-  return full_snapshot_name(date, rev, platform, hsh)
 
 # Main
